requirements.txt
streamlit
langchain==0.1.2
langchainhub
langchain-openai
openai
google-search-results
markdown
pydantic
Flask
requests
termcolor
python-dotenv
